Address the numeric system review
Pins the value grammars formally — IntegerString as 0|-?[1-9][0-9]* (one spelling per integer) and DecimalString as the JSON number grammar plus exact-case IEEE specials — and states the display precedence rule: a resolved display value wins for presentation, with the type's static facts as fallback. fixedPointTypeNode gains the display slot so static scale can compose with a contextual unit, its fixed-size rule now states its rationale (a fixed point presupposes a fixed bit width), ticksPerSecond narrows to u32 in line with the bounded-attributes policy, and the float display slot documents that only unit is meaningful. Also removes a duplicated test list entry and a stale 'number' in the injectable union docs.

+++ src/spec/nodes/typeNodes/DateTimeTypeNode.ts
@@ -1,11 +1,11 @@
-import { attribute, defineNode, node, optionalAttribute, u64 } from '../../../api';
+import { attribute, defineNode, node, optionalAttribute, u32 } from '../../../api';
 import { transformsAttribute } from '../transformNodes';
 import { examples } from './DateTimeTypeNode.examples';
 
 export const dateTimeTypeNode = defineNode('dateTimeTypeNode', {
     docs: ['A point in time encoded as an integer count of ticks since the Unix epoch.'],
     attributes: [
-        optionalAttribute('ticksPerSecond', u64(), {
+        optionalAttribute('ticksPerSecond', u32(), {
             docs: [
                 'How many ticks make one second. Defaults to `1` (the value is in seconds since the epoch).',
                 'Common choices are `1000` (milliseconds), `1000000` (microseconds), and `1000000000` (nanoseconds).',
