+++ src/spec/nodes/typeNodes/DurationTypeNode.ts
@@ -1,4 +1,4 @@
-import { attribute, defineNode, node, optionalAttribute, u64 } from '../../../api';
+import { attribute, defineNode, node, optionalAttribute, u32 } from '../../../api';
 import { transformsAttribute } from '../transformNodes';
 import { examples } from './DurationTypeNode.examples';
 
@@ -8,7 +8,7 @@ export const durationTypeNode = defineNode('durationTypeNode', {
         'Renderers typically format the value as `HH:mm:ss` or a coarser human-readable form.',
     ],
     attributes: [
-        optionalAttribute('ticksPerSecond', u64(), {
+        optionalAttribute('ticksPerSecond', u32(), {
             docs: [
                 'How many ticks make one second. Defaults to `1` (the value is already in seconds).',
                 'Common choices are `1000` (milliseconds), `1000000` (microseconds), and `1000000000` (nanoseconds).',
